Added Broadcasting Feature
Users in the 'SUDO_USERS' list can broadcast messages to the bot users. Also they can access the subs count. If the 'AUTH_USERS' list is empty, the bot can be used by publicly.

+++ plugins/thumbail.py
@@ -10,6 +10,7 @@ import asyncio
 from presets import Presets
 from pyrogram.types import Message
 from pyrogram import Client, filters
+from library.buttons import reply_markup_thumb, reply_markup_close
 
 
 if bool(os.environ.get("ENV", False)):
@@ -21,6 +22,9 @@ else:
 @Client.on_message(filters.private & filters.photo)
 async def save_photo(bot, m: Message):
     msg = await m.reply_text(Presets.WAIT_MESSAGE, reply_to_message_id=m.message_id)
+    if m.from_user.id in Config.SUDO_USERS:
+        await msg.edit(Presets.PROMPT_THUMB, reply_markup=reply_markup_thumb)
+        return
     if Config.AUTH_USERS and (m.from_user.id not in Config.AUTH_USERS):
         await m.delete()
         await msg.edit_text(Presets.NOT_AUTH_TXT)
@@ -29,6 +33,4 @@ async def save_photo(bot, m: Message):
         return
     thumb_image = os.getcwd() + "/" + "thumbnails" + "/" + str(m.from_user.id) + ".jpg"
     await bot.download_media(m, thumb_image)
-    await msg.edit_text(Presets.SAVED_THUMB)
-    await asyncio.sleep(5)
-    await msg.delete()
+    await msg.edit_text(Presets.SAVED_THUMB, reply_markup=reply_markup_close)
